requirements/dev.txt
djangorestframework
django-filter
pylint
pylint_django
drf-yasg